Replace implementation of IndentationRule with the new implementation and delete the old temporary class
